Refactoring based on CI
